requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
geopy==2.4.1
google-generativeai==0.5.4
pydantic==2.6.4
requests==2.31.0

# AI - LangChain
langchain>=1.0.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Database
sqlalchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1

# Background Tasks & Caching
celery==5.3.6
redis==5.0.3

# WebSockets
websockets==12.0
